Clean up config
